Fix oEmbed cache_age type Do not attempt to download metadata repeatedly if it has already failed once Fix "Other Websites" filter displaying an incorrect website name

+++ core/src/main/java/net/brlns/gdownloader/downloader/extractors/OEmbedMetadataExtractor.java
@@ -175,7 +175,7 @@ public class OEmbedMetadataExtractor implements IMetadataExtractor {
         @JsonProperty("provider_url")
         private String providerUrl;
         @JsonProperty("cache_age")
-        private Integer cacheAge;
+        private Long cacheAge;
         @JsonProperty("thumbnail_url")
         private String thumbnailUrl;
         @JsonProperty("thumbnail_width")
